baseline functional archive debugged
